% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/center_scale_xy.R
\name{center_scale_xy}
\alias{center_scale_xy}
\title{center_scale_xy}
\usage{
center_scale_xy(x, scale = T)
}
\arguments{
\item{x}{: dataframe

Function return the scaled datafrme with the mean and sd.}

\item{scale}{: a boolean}
}
\value{
mean : mean of dataframe x
sd   : sd of dataframe x
X    : the scaled dataframe
}
\description{
center_scale_xy
}
\examples{
data(iris)
x = iris[,1:4]
center_scale_xy(x,scale = T)


}
