% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circleFun.R
\name{circleFun}
\alias{circleFun}
\title{circleFunction}
\usage{
circleFun(center = c(0, 0), diameter = 1, npoints = 100)
}
\arguments{
\item{center}{: origine of circle}

\item{diameter}{: diameter of circle}

\item{npoints}{: number of points to create circle}
}
\value{
x = coordonate x of points
y = coordonate y of points
}
\description{
circleFunction
}
\examples{
circleFun(center = c(0,0),diameter = 1, npoints = 100)

}
