% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contribvar.PlsDA.R
\name{contribvar.PlsDA}
\alias{contribvar.PlsDA}
\title{Plot contribution variable}
\usage{
contribvar.PlsDA(object)
}
\arguments{
\item{object}{An object of class Pls-DA}
}
\value{
corrplot of contribution variables
}
\description{
return a graph of contribution variables.
}
\examples{
fit_launch : An object of class Pls-DA
contribvar.PlsDA(fit_launch)

}
