% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlationplot_plsDA.R
\name{correlationplot_plsDA}
\alias{correlationplot_plsDA}
\title{Correlation plot}
\usage{
correlationplot_plsDA(object, usedcomp = 1)
}
\arguments{
\item{object}{An object of class Pls-DA}

\item{usedcomp}{an integer for the number of componant

A graphic to show correlation between variables}
}
\value{
correlation plot of variables
}
\description{
Correlation plot
}
\examples{
fit_launch : an object of class Pls-DA
correlationplot_plsDA(fit_launch, usedcomp=1)

}
