% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cos2.PlsDA.R
\name{Cos2.PlsDA}
\alias{Cos2.PlsDA}
\title{Cos2 graph}
\usage{
Cos2.PlsDA(object)
}
\arguments{
\item{object}{An object of class Pls-DA}
}
\value{
Cos2 : quality representation of variables
}
\description{
Cos2 graph
}
\examples{
fit_launch : An object of class Pls-DA
Cos2.PlsDA(fit_launch)
}
