% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit}
\alias{fit}
\title{Title}
\usage{
fit(
  formula,
  data,
  ncomp = NULL,
  algorithm = "simpls",
  var.select = F,
  threshold.algorithm = 0.001,
  threshold.vip = 1,
  iter.max = 100,
  nfold = NULL
)
}
\arguments{
\item{nfold}{}
}
\description{
Title
}
