% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dummies.R
\name{get_dummies}
\alias{get_dummies}
\title{Function Get Dummies}
\usage{
get_dummies(d)
}
\arguments{
\item{X}{dataframe}
}
\value{
A dataframe of variables binarize
}
\description{
It's a function who take a matrix or dataframe of one qualitative variable and to return variables binarize
}
\examples{
data(iris)
y = iris[,5]
y = get_dummies(y)
print(y)
}
