% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_first_singular_vectors_svd.R
\name{get_first_singular_vectors_svd}
\alias{get_first_singular_vectors_svd}
\title{get_first_singular_vectors_svd}
\usage{
get_first_singular_vectors_svd(X, Y)
}
\arguments{
\item{X}{variables features}

\item{Y}{variable to explain}
}
\value{
u, v
}
\description{
Return the first left and right singular vectors of X'Y
}
\examples{
data(iris)
X = iris[,1:4]
Y = iris[,5]
get_svd = get_first_singular_vectors_svd(X, Y)
get_svd$u
get_svd$v

}
