% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotInd.PlsDA.R
\name{plotInd.PlsDA}
\alias{plotInd.PlsDA}
\title{Individu plot}
\usage{
plotInd.PlsDA(object, comp1 = 1, comp2 = 2)
}
\arguments{
\item{object}{An object of class Pls-DA}

\item{comp1}{choose the first componant}

\item{comp2}{choose the second componant}
}
\value{
p : Plot of individu
}
\description{
Create the individu plot on plan with two componants.
}
\examples{
fit_launch : an object of class Pls-DA
plotInd.PlsDA(fit_launch, comp1 = 1, comp2 = 2)

}
