% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotVar.PlsDA.R
\name{plotVar.PlsDA}
\alias{plotVar.PlsDA}
\title{plotVar.PlsDA}
\usage{
plotVar.PlsDA(object, comp1 = 1, comp2 = 2)
}
\arguments{
\item{object}{An object of class Pls-DA}

\item{comp1}{choose the first componant}

\item{comp2}{choose the second componant}
}
\value{
A correlation circle with coordinate of variables
}
\description{
Create a correlation circle with coordinate of variables.
}
\examples{
fit_launch : an object of class Pls-DA
plotVar.PlsDA(fit_launch)

}
