% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlsDAcv.R
\name{PlsDA.cv}
\alias{PlsDA.cv}
\title{Cross-validation of PLSDA model}
\usage{
PlsDA.cv(X, Y, threshold = 0.001, nfold = NULL)
}
\arguments{
\item{X}{the numeric data frame or matrix of observations. Missing values are not allowed.}

\item{Y}{a vector or matrix of responses. Missing values are not allowed.}

\item{threshold}{the threshold used in NIPALS algorithm. Default to 0.001.}

\item{nfold}{the number of folds used in cross validation. Default to NULL. If nfold is missing, the number of folds is the minimum of rows’ or columns’ number of data.}
}
\value{
Q2 :	the coefficient of “leave-one-out” cross-validation. If Q2 >= 0.05, we take the component.

ncomp	: the number of components obtained by “leave-one-out” cross-validation.
}
\description{
A “leave-one-out” cross-validation function used in fit.
}
\details{
This function performs a “leave-one-out” cross-validation based on NIPALS algorithm on a model fit. When ncomp = NULL, the number of components is obtained by cross-validation. When a number of components is specified, cross-validation results are calculated for each component.

Leave-one-out cross-validation uses the following approach to evaluate a model:

- Split a dataset into a training set and a test set, using all but one observation as part of the training set (we only leave one observation out from the training set).

- Build the model using only data from the training set.

- Use the model to predict the response value of the one observation left out of the model and calculate Q2 coefficient.

- Repeat the process nfold times.
}
\examples{
data(iris)
PlsDA.cv(iris[,1:4], iris[,5])

}
