% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlsDAvip.R
\name{PlsDA.vip}
\alias{PlsDA.vip}
\title{PlsDA.vip}
\usage{
PlsDA.vip(X, Y, ncomp, threshold = 1, algorithm = "simpls")
}
\arguments{
\item{X}{a dataframe with features}

\item{Y}{a dataframe with variable to explain}

\item{ncomp}{a integer}

\item{threshold}{a integer}

\item{algorithm}{Type of algorithm to choose for selection}
}
\value{
VIP : A dataframe

newX : A new dataframe with features choose

important.variables : list of important variables

threshold : An integer

ncomp : An integer, number of componants
}
\description{
PlsDA.vip
}
\examples{
data(iris)
Y = iris[,5]
X = iris[,1:4]
PlsDA.vip(X, Y, ncomp = 2,algorithm="simpls")


}
