% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_PlsDA.R
\name{print.PLSDA}
\alias{print.PLSDA}
\title{Function print}
\usage{
\method{print}{PLSDA}(objet)
}
\arguments{
\item{objet}{An object of class Pls-DA}
}
\value{
classement : A dataframe
}
\description{
A function who is surchage by the class Pls-DA to print the classification function.
}
\examples{
fit_launch : an object of class Pls-DA
print.PLSDA(fit_launch)

}
