% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screeplot_plsDA.R
\name{screeplot_plsDA}
\alias{screeplot_plsDA}
\title{Screeplot}
\usage{
screeplot_plsDA(
  object,
  colorbar = NULL,
  linecolor = "black",
  marcker = "black"
)
}
\arguments{
\item{object}{: An object of class Pls-DA}

\item{linecolor}{: colorline}

\item{marcker}{: color point line}

\item{cololbar}{: color of barplot}
}
\value{
PVEplot : a screeplot to determinded the number of componants with the criteria of elbow method
}
\description{
Screeplot
}
\examples{
fit_launche : An object of class Pls-DA
screeplot_plsDA(object,barfill="steelblue",linecolor = "black")

}
