% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIMPLS.R
\name{SIMPLS}
\alias{SIMPLS}
\title{SIMPLS algorithm}
\usage{
SIMPLS(X, Y, n_components = ncomp)
}
\arguments{
\item{X}{variables features}

\item{Y}{variable to explain}

\item{n_components}{number of composants}
}
\value{
X : dataframe of variables features

Y : datafrme of variable to explain

classification : classification report (coefficient, intercept)

Coefficients : coefficients of regression

Intercept: intercept of regression

x_weights : dataframe weights of X

y_weights : dataframe weights of y

Scores_X : dataframe scores of X

Scores_Y : dataframe scores of X

Loadings_X : dataframe loadings of X

Loadings_Y : dataframe loadings of Y

R2X : R square of X

R2Y : R square of Y

R2Var : R square of variables
}
\description{
SIMPLS algorithm
}
\examples{
data(iris)
X = iris[,1:4]
Y = iris[,5]
n_components = 2
SIMPLS(X,Y,n_components)

}
