% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.plsDA.R
\name{summary.plsDA}
\alias{summary.plsDA}
\title{summary.plsDA}
\usage{
\method{summary}{plsDA}(object)
}
\arguments{
\item{object}{An object of class Pls-DA}
}
\value{
Print summary of results for the algorithms nipals and simpls.
}
\description{
Print the summary of results of PLS DA package.
}
\examples{
fit_launch : an object of class Pls-DA
summary.plsDA(fit_launch)

}
