% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svd_flip_1d.R
\name{svd_flip_1d}
\alias{svd_flip_1d}
\title{svd_flip_1d}
\usage{
svd_flip_1d(u, v)
}
\arguments{
\item{u}{first left singular vector of X'Y}

\item{v}{first right singular vector of X'Y}
}
\value{
U, V modified
}
\description{
svd_flip would force us to convert to 2d array and would also return 2d arrays.
}
\details{
We don't want that.
}
\examples{
get_svd = svd_flip_1d(u, v)
u = get_svd$u
v = get_svd$v

svd_flip_1d = function(u, v)
}
