% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_test_split.R
\name{train_test_split}
\alias{train_test_split}
\title{train_test_split}
\usage{
train_test_split(dataframe)
}
\arguments{
\item{dataframe}{A dataframe}
}
\value{
Dataframe : A dataframe
Df_Train  : Dataframe split into train dataset
Df_Test   : Dataframe split into test dataset
}
\description{
Split dataframe into Train and test dataset.
}
\examples{
data(iris)
train_test_split(iris)

}
