% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var.coord.R
\name{var.coord}
\alias{var.coord}
\title{Coordonates variables}
\usage{
var.coord(object)
}
\arguments{
\item{object}{An object of class Pls-DA}
}
\value{
var.coord : coordinates of variables
}
\description{
Coordonates variables
}
\examples{
fit_launch : an object of class Pls-DA

var.coord(fit_launch)

}
