% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.vip.R
\name{plot.vip}
\alias{plot.vip}
\title{Plot VIP}
\usage{
\method{plot}{vip}(ObjectVIP)
}
\arguments{
\item{ObjectVIP}{An object of class Pls-DA}
}
\value{
A plot VIP to select variable in the modele
}
\description{
Plot VIP
}
\examples{
fit_launch : An object of class Pls-DA
plot.vip(fit_launch)

}
