% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screeplot_plsDA.R
\name{screeplot_plsDA}
\alias{screeplot_plsDA}
\title{screeplot_plsDA}
\usage{
screeplot_plsDA(object, barfill = "steelblue", linecolor = "black")
}
\arguments{
\item{object}{: An object of class Pls-DA}

\item{barfill}{: color barplot}

\item{linecolor}{: color line}
}
\value{
PVEplot : a screeplot to determinded the number of componants with the criteria of elbow method
}
\description{
screeplot_plsDA
}
\examples{
fit_launche : An object of class Pls-DA
screeplot_plsDA(object,barfill="steelblue",linecolor = "black")

}
